Refactor GrAITests.
