+++ Tests/GrAITests/Base/Input1DMSE1DCase.swift
@@ -9,14 +9,97 @@ import XCTest
 import GrAIdient
 import GrAITestsUtils
 
-let NB_RETRY = 3
-
 ///
 /// A class that will test a model with a structural hypothesis:
 /// the model last layer is a MSE1D layer, the model first layer is a Input1D.
 ///
-class Input1DMSE1DCase: MSE1DCase
+class Input1DMSE1DCase: XCTestCase, IOCase
 {
+    /// Batch size of data.
+    var batchSize: Int = -1
+    /// Optimizer parameters.
+    var optimizerParams = GrAI.Optimizer.Params()
+    
+    /// Systematic call before test begins.
+    override func setUp()
+    {
+        batchSize = 5
+        _ = MetalKernel.get
+        GrAI.Opti.GPU = true
+        
+        setOptimizerParams(params: &optimizerParams)
+        optimizerParams.nbLoops = 3
+    }
+    
+    ///
+    /// A function to create/set ground truth to the model.
+    ///
+    /// - Parameters:
+    ///     - groundTruth: The ground truth to set.
+    ///     - model: The model.
+    /// - Returns: The ground truth.
+    ///
+    func setLoss(_ groundTruth: [[Double]]?, _ model: Model) -> [[Double]]
+    {
+        let lastLayer = model.layers.last as! MSE1D
+        let gt: [[Double]]
+        if let groundTruthTmp = groundTruth
+        {
+            gt = groundTruthTmp
+        }
+        else
+        {
+            gt = buildData(dim1: getBatchSize(model), dim2: 1)
+        }
+        
+        if GrAI.Opti.GPU
+        {
+            try! lastLayer.lossDerivativeGPU(gt)
+        }
+        else
+        {
+            try! lastLayer.lossDerivativeCPU(gt)
+        }
+        return gt
+    }
+    
+    ///
+    /// A function to get loss of a model.
+    ///
+    /// - Parameters:
+    ///     - groundTruth: The ground truth to set.
+    ///     - model: The model.
+    /// - Returns: The loss value.
+    ///
+    func getLoss(_ groundTruth: [[Double]], _ model: Model) -> Double
+    {
+        let lastLayer = model.layers.last as! MSE1D
+        if GrAI.Opti.GPU
+        {
+            return Double(try! lastLayer.getLossGPU(groundTruth))
+        }
+        else
+        {
+            return try! lastLayer.getLossCPU(groundTruth)
+        }
+    }
+    
+    ///
+    /// A function to get the gradients of weights approximations..
+    ///
+    /// - Parameters:
+    ///     - groundTruth: The ground truth.
+    ///     - model: The model.
+    /// - Returns: The gradients of weights approximations.
+    ///
+    func getGradientsApprox(
+        _ groundTruth: [[Double]],
+        _ model: Model) -> [Double]
+    {
+        let lastLayer = model.layers.last as! MSE1D
+        return try! lastLayer.collectGradientsApprox(groundTruth)
+    }
+    
     ///
     /// A function to create/set data to the model.
     ///
@@ -86,330 +169,11 @@ class Input1DMSE1DCase: MSE1DCase
         return modelNew
     }
     
-    ///
-    /// Run Gradient Checking test.
-    ///
-    /// The goal is to compare the gradients of weights that are computed through `backward`
-    /// to an estimation that is being computed through `forwardGC`.
-    ///
-    /// - Parameters:
-    ///     - trainer: The testing pipeline to run.
-    ///     - nbRetry: The maximum number we can retry the test.
-    ///     - diffThreshold: The threshold above which the relative difference is too high.
-    ///
-    func run(
-        _ trainer: GradTrainer,
-        nbRetry: Int = NB_RETRY,
-        diffThreshold: Double = 0.000001)
+    /// A list of functions that transform the model into another one.
+    var transforms: [(Model) -> Model]
     {
-        let model = trainer.model!
-        let lastLayer = model.layers.last as! MSE1D
-        let layersGraph = model.getGraph(lastLayer)
-        
-        retryNumeric(
-            nbRetry: nbRetry,
-            {
-                () throws in
-                try trainer.run(
-                    layersGraph: layersGraph,
-                    setData: self.setData,
-                    setLoss: self.setLoss,
-                    getGradientsApprox: self.getGradientsApprox)
-                {
-                    (gradDiff: Double) in
-                    if gradDiff > diffThreshold
-                    {
-                        throw TestError.Numeric
-                    }
-                }
-            },
-            {
-                () in
-                XCTAssert(false)
-            }
-        )
-    }
-    
-    ///
-    /// Run Flow test.
-    ///
-    /// The goal is to compare the gradients of weights computed in the CPU execution context with
-    /// the gradients of weights computed in the GPU execution context.
-    ///
-    /// - Parameters:
-    ///     - trainer: The testing pipeline to run.
-    ///     - nbRetry: The maximum number we can retry the test.
-    ///     - diffThreshold: The threshold above which the relative difference is too high.
-    ///
-    func run(
-        _ trainer: FlowTrainer,
-        nbRetry: Int = NB_RETRY,
-        diffThreshold: Double = 0.000001)
-    {
-        retryNumeric(
-            nbRetry: nbRetry,
-            {
-                () throws in
-                try trainer.run(
-                    setData: self.setData,
-                    setLoss: self.setLoss)
-                {
-                    (gradDiff: Double) in
-                    if gradDiff > diffThreshold
-                    {
-                        throw TestError.Numeric
-                    }
-                }
-            },
-            {
-                () in
-                XCTAssert(false)
-            }
-        )
-    }
-    
-    ///
-    /// Run Flow Reset test.
-    ///
-    /// The goal is to compare the gradients of weights computed in the CPU execution context with
-    /// the gradients of weights computed in the GPU execution context.
-    ///
-    /// - Parameters:
-    ///     - trainer: The testing pipeline to run.
-    ///     - nbRetry: The maximum number we can retry the test.
-    ///     - diffThreshold: The threshold above which the relative difference is too high.
-    ///
-    func run(
-        _ trainer: FlowResetTrainer,
-        nbRetry: Int = NB_RETRY,
-        diffThreshold: Double = 0.000001)
-    {
-        retryNumeric(
-            nbRetry: nbRetry,
-            {
-                () throws in
-                try trainer.run(
-                    setData: self.setData,
-                    setLoss: self.setLoss)
-                {
-                    (gradDiff: Double) in
-                    if gradDiff > diffThreshold
-                    {
-                        throw TestError.Numeric
-                    }
-                }
-            },
-            {
-                () in
-                XCTAssert(false)
-            }
-        )
-    }
-    
-    ///
-    /// Run Flow Reverse test.
-    ///
-    /// The goal is to compare the gradients of weights computed in the CPU execution context with
-    /// the gradients of weights computed in the GPU execution context.
-    ///
-    /// - Parameters:
-    ///     - trainer: The testing pipeline to run.
-    ///     - nbRetry: The maximum number we can retry the test.
-    ///     - diffThreshold: The threshold above which the relative difference is too high.
-    ///
-    func run(
-        _ trainer: FlowReverseTrainer,
-        nbRetry: Int = NB_RETRY,
-        diffThreshold: Double = 0.000001)
-    {
-        retryNumeric(
-            nbRetry: nbRetry,
-            {
-                () throws in
-                try trainer.run(
-                    setData: self.setData,
-                    setLoss: self.setLoss)
-                {
-                    (gradDiff: Double) in
-                    if gradDiff > diffThreshold
-                    {
-                        throw TestError.Numeric
-                    }
-                }
-            },
-            {
-                () in
-                XCTAssert(false)
-            }
-        )
-    }
-    
-    ///
-    /// Run Inference test.
-    ///
-    /// The goal is to compare the losses computed in the CPU execution context with
-    /// the losses computed in the GPU execution context during the inference phase.
-    ///
-    /// - Parameters:
-    ///     - trainer: The testing pipeline to run.
-    ///     - nbRetry: The maximum number we can retry the test.
-    ///     - diffThreshold: The threshold above which the relative difference is too high.
-    ///
-    func run(
-        _ trainer: InferenceTrainer,
-        nbRetry: Int = NB_RETRY,
-        diffThreshold: Double = 0.001)
-    {
-        retryNumeric(
-            nbRetry: nbRetry,
-            {
-                () throws in
-                try trainer.run(
-                    setData: self.setData,
-                    setLoss: self.setLoss,
-                    getLoss: self.getLoss)
-                {
-                    (lossDiff: Double) in
-                    if lossDiff > diffThreshold
-                    {
-                        throw TestError.Numeric
-                    }
-                }
-            },
-            {
-                () in
-                XCTAssert(false)
-            }
-        )
-    }
-    
-    ///
-    /// Run Loading test.
-    ///
-    /// The goal is to compare the losses computed in the CPU execution after havinng loaded the
-    /// model from the disk and do the same in the GPU execution context.
-    ///
-    /// - Parameters:
-    ///     - trainer: The testing pipeline to run.
-    ///     - nbRetry: The maximum number we can retry the test.
-    ///     - diffThreshold: The threshold above which the relative difference is too high.
-    ///
-    func run(
-        _ trainer: LoadTrainer,
-        nbRetry: Int = NB_RETRY,
-        diffThreshold: Double = 0.001)
-    {
-        retryNumeric(
-            nbRetry: nbRetry,
-            {
-                () throws in
-                try trainer.run(
-                    setData: self.setData,
-                    setLoss: self.setLoss,
-                    getLoss: self.getLoss)
-                {
-                    (diffCPU: Double, diffGPU: Double) in
-                    if diffCPU > diffThreshold
-                    {
-                        throw TestError.Numeric
-                    }
-                    if diffGPU > diffThreshold
-                    {
-                        throw TestError.Numeric
-                    }
-                }
-            },
-            {
-                () in
-                XCTAssert(false)
-            }
-        )
-    }
-    
-    ///
-    /// Run Transform tests.
-    ///
-    /// The goal is to compare the losses computed in the CPU execution
-    /// after transforming the model and do the same in the GPU execution context.
-    ///
-    /// - Parameters:
-    ///     - trainer: The testing pipeline to run.
-    ///     - nbRetry: The maximum number we can retry the test.
-    ///     - diffThreshold: The threshold above which the relative difference is too high.
-    ///
-    func run(
-        _ trainer: TransformTrainer,
-        nbRetry: Int = NB_RETRY,
-        diffThreshold: Double = 0.001)
-    {
-        retryNumeric(
-            nbRetry: nbRetry,
-            {
-                () throws in
-                try trainer.run(
-                    transforms: [self.copy, self.copyInPlace],
-                    setData: self.setData,
-                    setLoss: self.setLoss,
-                    getLoss: self.getLoss)
-                {
-                    (diffCPU: Double, diffGPU: Double) in
-                    if diffCPU > diffThreshold
-                    {
-                        throw TestError.Numeric
-                    }
-                    if diffGPU > diffThreshold
-                    {
-                        throw TestError.Numeric
-                    }
-                }
-            },
-            {
-                () in
-                XCTAssert(false)
-            }
-        )
-    }
-    
-    ///
-    /// Run Clipping test.
-    ///
-    /// The goal is to compare the norm of the gradients of the weights with a threshold.
-    ///
-    /// - Parameters:
-    ///     - trainer: The testing pipeline to run.
-    ///     - nbRetry: The maximum number we can retry the test.
-    ///     - diffThreshold: The threshold above which the relative difference is too high.
-    ///     - normClipping: The threshold above which gradients must be cut.
-    ///
-    func run(
-        _ trainer: NormTrainer,
-        nbRetry: Int = NB_RETRY,
-        diffThreshold: Double = 0.001,
-        normClipping: Double = 0.001)
-    {
-        optimizerParams.gradientClipping = true
-        optimizerParams.normThreshold = normClipping
-        trainer.optimizerParams = optimizerParams
-        
-        retryNumeric(
-            nbRetry: nbRetry,
-            {
-                () throws in
-                try trainer.run(
-                    setData: self.setData,
-                    setLoss: self.setLoss)
-                {
-                    (normDiff: Double) throws in
-                    if normDiff > diffThreshold
-                    {
-                        throw TestError.Numeric
-                    }
-                }
-            },
-            {
-                () in
-                XCTAssert(false)
-            }
-        )
+        get {
+            return [copy, copyInPlace]
+        }
     }
 }
